Add HTML repr for Attachment
