+++ src/scitacean/_html_repr/styles/attachment.css
@@ -1,19 +1,3 @@
-.cean-attachment {
-    display: table;
-    clear: both
-}
-
-.cean-attachment-info {
-    float: left;
-    width: 75%
-}
-
-.cean-thumbnail {
-    float: left;
-    width: 20%;
-    margin-left: 5%;
-}
-
 tr.cean-header {
     border-bottom: 2px solid var(--jp-layout-color3);
 }
@@ -23,20 +7,17 @@ tr.cean-header {
     background: var(--jp-layout-color0);
 }
 
-.cean-info-line {
+.cean-attachment .cean-info-line {
     /* match the margin of the tables */
     margin-left: 1em;
-}
-
-.cean-dataset-type {
     font-weight: bold;
 }
 
-.cean-dataset th {
+.cean-attachment th {
     text-align: left !important;
 }
 
-.cean-dataset td {
+.cean-attachment td {
     white-space: nowrap !important;
     overflow: hidden;
     text-overflow: ellipsis;
@@ -88,13 +69,7 @@ td.cean-field-description {
     color: var(--jp-content-font-color2);
 }
 
-.cean-missing-value,
-.cean-error {
-    border-left: var(--jp-error-color1) 4px solid !important;
-    border-right: var(--jp-error-color1) 4px solid !important;
-}
-
-table.cean-dataset {
+table.cean-attachment-table {
     /* space between details marker and table */
     margin-left: 1em !important;
     /* remove space between tables */
@@ -105,66 +80,19 @@ table.cean-dataset {
     width: 100%;
 }
 
-.cean-dataset-details table.cean-dataset {
-    /* Jupyter's default space after a table */
-    margin-bottom: 1em;
-
-    border-top: 2px solid var(--jp-layout-color3);
-}
-
 /* Position the summary table directly under the regular table */
-details[open].cean-dataset-details summary {
+details[open].cean-attachment-details summary {
     height: 0;
     padding: 0;
 }
 
 /* hide the summary when the details are open */
-details[open].cean-dataset-details summary span {
+details[open].cean-attachment-details summary span {
     visibility: hidden;
 }
 
 details summary {
     cursor: pointer;
-}
-
-.cean-file-info {
-    padding-left: 1em;
-}
-
-.cean-archived-file-info {
-    padding-left: 4em;
-}
-
-.cean-file-info-size {
-    color: var(--jp-content-font-color1);
-}
-
-table.cean-files,
-table.cean-metadata {
-    /* match spacing of dataset tables */
-    margin-left: 1em !important;
-    width: 100%;
-}
-
-.cean-file-local-path,
-.cean-file-remote-path {
-    width: 40%;
-    text-align: left !important;
-}
-
-.cean-file-size {
-    width: 10%;
-    text-align: left !important;
-}
-
-.cean-metadata-name {
-    width: 30%;
-    text-align: left !important;
-}
-
-.cean-metadata-value {
-    width: 70%;
-    text-align: left !important;
 }
 
 .cean-lock {
@@ -173,8 +101,8 @@ table.cean-metadata {
 }
 
 .cean-lock svg {
-    max-width: 100%;
-    height: auto;
+  max-width: 100%;
+  height: auto;
 }
 
 .cean-lock path {
